pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "subtimizer"
version = "1.0.0"
description = "A Computational Workflow for Structure-Guided Design of Potent and Selective Kinase Peptide Substrates"
readme = "README.md"
authors = [
  { name = "Abeeb Yekeen", email = "Abeeb.Yekeen@UTSouthwestern.edu" },
]
requires-python = ">=3.9"
dependencies = [
    "biopython",
    "pandas",
    "numpy",
    "matplotlib",
    "seaborn",
    "colorcet",
    "click",  # For CLI
    "weblogo",
    "scipy", 
]
classifiers = [
    "Programming Language :: Python :: 3",
]
license = "MIT"
license-files = ["LICENSE"]

[project.urls]
Homepage = "https://github.com/abeebyekeen/subtimizer"
Publication = "https://doi.org/10.1101/2025.07.04.663216"

[project.scripts]
subtimizer = "subtimizer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/subtimizer"]
